New icons.zip
This includes Daves' new skyanytime icon- well done m8! trutvplus1 was nuked a day or two ago.
New strategy for 2017: when such channels die, is it for good, or temp (unpaid bills)? We have no
way of knowing... So we'll take the dead icon out and SAVE it in case the channel comes back!

We monitor these kind of sheenanigams every morning. If it reappears, we'll add the icon back!